pip-requirements.txt
boto3
hurry.filesize
prompt_toolkit